requirements.txt
langgraph
streamlit
pydantic-ai
logfire
html2text
beautifulsoup4
playwright
python-dotenv
uuid
aiohttp 